PAY-3649 Code review fixes.
